+++ ChangeLog.md
@@ -27,6 +27,10 @@ Incoming in 0.10.0
     * Restored `PatternBuilder::prepare()`, but without alteration. #78
     * Add `Match.tail()`. #83
     * Add `Match.byteTail()`. #83
+    * Add `MatchGroup.tail()`. #83
+    * Add `MatchGroup.byteTail()`. #83
+    * Add `ReplaceMatchGroup.tail()`. #83
+    * Add `ReplaceMatchGroup.byteTail()`. #83
 
 Added in 0.9.10
 ---------------